myapp-api/pyproject.toml
[tool.poetry]
name = "myapp-api"
version = "0.1.0"
description = "Rest API for MyApp"
authors = ["Yin-Chi Chan <ycc39@cam.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
myapp-models = { version = "^0.1.0", source = "local" }
fastapi = { extras = ["standard"], version = "^0.112.0" }
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
SQLAlchemy = "^2.0.32"
sqlmodel = "^0.0.21"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pylint = "^3.2.6"


[[tool.poetry.source]]
name = "local"
url = "http://host.docker.internal:31415/simple"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 100
color_output = true
known_first_party = ["models"]

[tool.autopep8]
max_line_length = 100
